requirements.txt
Flask==1.0.2
mysql-connector-python
cffi
paramiko
sshtunnel
pyyaml
ua-parser
user-agents
